+++ app/src/androidTest/java/com/andela/art/serialentry/securitydashboard/OkHttpIdlingResourceRule.java
@@ -1,4 +1,4 @@
-package com.andela.art.serialentry.presentation;
+package com.andela.art.serialentry.securitydashboard;
 
 import android.support.test.espresso.IdlingRegistry;
 import android.support.test.espresso.IdlingResource;
