+++ app/src/main/AndroidManifest.xml
@@ -22,7 +22,7 @@
             </intent-filter>
         </activity>
 
-        <activity android:name=".securitydashboard.SecurityDashboardActivity" />
+        <activity android:name=".securitydashboard.presentation.SecurityDashboardActivity" />
         <activity android:name=".settings.SettingsActivity"></activity>
