+++ app/src/main/java/com/andela/art/login/LoginActivity.java
@@ -13,7 +13,7 @@ import android.view.WindowManager;
 import android.widget.Toast;
 
 import com.andela.art.R;
-import com.andela.art.securitydashboard.SecurityDashboardActivity;
+import com.andela.art.securitydashboard.presentation.SecurityDashboardActivity;
 import com.andela.art.databinding.ActivityLoginBinding;
 import com.google.android.gms.auth.api.signin.GoogleSignIn;
 import com.google.android.gms.auth.api.signin.GoogleSignInAccount;
