+++ app/src/main/res/layout/security_dashboard.xml
@@ -1,53 +1,52 @@
-<?xml version="1.0" encoding="utf-8"?>
-<android.support.constraint.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<layout
+    xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <android.support.constraint.ConstraintLayout
     xmlns:tools="http://schemas.android.com/tools"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/primaryBalticSea"
-    tools:context="com.andela.art.securitydashboard.SecurityDashboardActivity">
+    tools:context="com.andela.art.securitydashboard.presentation.SecurityDashboardActivity">
 
     <include
         android:id="@+id/mToolBar"
         layout="@layout/tool_bar"/>
 
     <de.hdodenhof.circleimageview.CircleImageView
-        android:id="@+id/profile_picture"
+        android:id="@+id/profilePhoto"
         android:layout_width="@dimen/profile_image_diameter"
         android:layout_height="@dimen/profile_image_diameter"
         android:layout_marginTop="@dimen/profile_margin_top"
-        tools:src="@mipmap/ic_launcher_round"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/mToolBar"/>
 
     <TextView
-        android:id="@+id/full_name"
+        android:id="@+id/display_name"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginTop="@dimen/profile_margin_top"
-        tools:text="@string/full_name"
         android:textColor="@color/profileTextViewColor"
         android:fontFamily="@font/avenir_heavy"
         android:textSize="@dimen/twenty_two_sp"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/profile_picture" />
+        app:layout_constraintTop_toBottomOf="@+id/profilePhoto" />
 
     <TextView
         android:id="@+id/email_address"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        tools:text="@string/email"
         android:textColor="@color/profileTextViewColor"
         android:textSize="@dimen/twelve_sp"
         android:fontFamily="@font/avenir_book"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/full_name" />
+        app:layout_constraintTop_toBottomOf="@+id/display_name" />
 
     <Button
-        android:id="@+id/check_serial"
+        android:id="@+id/addSerial"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginTop="@dimen/profile_button_margin_top"
@@ -56,7 +55,6 @@
         android:paddingLeft="@dimen/profile_button_padding_left"
         android:paddingRight="@dimen/profile_button_padding_left"
         android:paddingTop="@dimen/profile_button_padding_top"
-        android:text="@string/check_serial"
         android:textColor="#ffffff"
         android:fontFamily="@font/avenir_book"
         android:textSize="@dimen/sixteen_sp"
@@ -64,6 +62,5 @@
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/email_address" />
-
-
 </android.support.constraint.ConstraintLayout>
+</layout>
