+++ app/src/mock/java/com/andela/art/securitydashboard/injection/FirebasePresenterModule.java
@@ -1,10 +1,10 @@
-package com.andela.art.serialentry.injection;
+package com.andela.art.securitydashboard.injection;
 
+import com.andela.art.securitydashboard.presentation.FirebasePresenter;
 import com.google.firebase.auth.FirebaseAuth;
 
 import dagger.Module;
 import dagger.Provides;
-import com.andela.art.serialentry.presentation.FirebasePresenter;
 
 /**
  * Created by zack on 3/29/18.
